@@ -24,13 +24,13 @@
         }
     },
     "require": {
-        "php": "^8.0",
-        "emartech/escher": "^3.0",
+        "php": ">=8.1",
+        "emartech/escher": "^4.0",
         "guzzlehttp/guzzle": "^7.4",
         "psr/log": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.4",
+        "phpunit/phpunit": "^10.0",
         "squizlabs/php_codesniffer": "^3.3"
     },
     "suggest": {
