@@ -4,16 +4,7 @@ namespace SessionValidator\Cache;
 
 interface CacheInterface
 {
-    /**
-     * @param string $key
-     * @return mixed
-     */
-    public function get($key);
+    public function get(string $key): mixed;
 
-    /**
-     * @param string $key
-     * @param mixed $value
-     * @return bool
-     */
-    public function set($key, $value);
+    public function set(string $key, mixed $value): bool;
 }
